[T6] company.list + getWithDeclarations ignorent les annulées + status to_complete (#3424)
