+++ packages/app/src/modules/domain/shared/declarationStatus.ts
@@ -8,13 +8,20 @@ export function isCancelled(declaration: {
 
 export function computeDeclarationStatus(
 	declaration:
-		| { status: string | null; currentStep: number | null }
+		| {
+				status: string | null;
+				currentStep: number | null;
+				cancelledAt?: Date | null;
+		  }
 		| undefined,
 ): DeclarationStatus {
 	if (
 		!declaration ||
 		(declaration.status === "draft" && (declaration.currentStep ?? 0) === 0)
 	) {
+		return "to_complete";
+	}
+	if (declaration.cancelledAt != null) {
 		return "to_complete";
 	}
 	if (declaration.status === "submitted") {
